requirements.txt
asgiref==3.7.1
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
mysqlclient==2.1.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
